+++ gear/csv_center_splitter/src/python/csv_center_splitter_app/run.py
@@ -19,7 +19,7 @@ from notifications.email_list import (
     EmailListError,
     get_redcap_email_list_client,
 )
-from outputs.errors import ListErrorWriter
+from outputs.error_writer import ListErrorWriter
 from utils.utils import parse_string_to_list
 
 from csv_center_splitter_app.main import run
@@ -83,8 +83,10 @@ class CSVCenterSplitterVisitor(GearExecutionEnvironment):
         target_project = context.config.get("target_project", None)
         staging_project_id = context.config.get("staging_project_id", None)
 
-        if not (target_project or staging_project_id):
-            raise GearExecutionError("No target or staging project provided")
+        if not target_project:
+            raise GearExecutionError("No target project provided")
+        if not staging_project_id:
+            raise GearExecutionError("No staging project provided")
 
         adcid_key = context.config.get("adcid_key", None)
         if not adcid_key: