+++ gear/form_transformer/src/python/form_csv_app/main.py
@@ -5,23 +5,24 @@ from collections import defaultdict
 from typing import Any, DefaultDict, Dict, List, MutableMapping, Optional, TextIO
 
 from configs.ingest_configs import ErrorLogTemplate, ModuleConfigs
+from flywheel.models.file_entry import FileEntry
 from flywheel.rest import ApiException
 from flywheel_adaptor.flywheel_proxy import ProjectAdaptor
 from inputs.csv_reader import CSVVisitor, read_csv
 from keys.keys import FieldNames, PreprocessingChecks, SysErrorCodes
+from outputs.error_logger import update_error_log_and_qc_metadata
+from outputs.error_models import FileQCModel, QCStatus
+from outputs.error_writer import ListErrorWriter
 from outputs.errors import (
-    ListErrorWriter,
     empty_field_error,
-    get_error_log_name,
     missing_field_error,
     partially_failed_file_error,
     preprocess_errors,
     preprocessing_error,
-    system_error,
     unexpected_value_error,
-    update_error_log_and_qc_metadata,
 )
 from preprocess.preprocessor import FormPreprocessor
+from pydantic import ValidationError
 from transform.transformer import BaseRecordTransformer, TransformerFactory
 from uploads.uploader import FormJSONUploader
 
@@ -118,7 +119,6 @@ class CSVTransformVisitor(CSVVisitor):
         Returns:
           True if the row was processed without error, False otherwise
         """
-
         self.__error_writer.clear()
 
         found_all = True
@@ -158,9 +158,7 @@ class CSVTransformVisitor(CSVVisitor):
             self.__module_configs.preprocess_checks
             and PreprocessingChecks.DUPLICATE_RECORD
             in self.__module_configs.preprocess_checks
-            and self.__preprocessor.is_existing_visit(
-                input_record=transformed_row, module=self.module
-            )
+            and self.__preprocessor.is_existing_visit(input_record=transformed_row)
         ):
             self.__existing_visits[subject_lbl].append(transformed_row)
             return True
@@ -267,7 +265,6 @@ class CSVTransformVisitor(CSVVisitor):
 
                 if not self.__preprocessor.preprocess(
                     input_record=transformed_row,
-                    module=self.module,
                     line_num=line_num,
                     ivp_record=ivp_packet,
                 ):
@@ -373,13 +370,13 @@ class CSVTransformVisitor(CSVVisitor):
             )
             return None
 
-        error_log_name = get_error_log_name(
-            module=self.module,
-            input_data=input_record,
-            errorlog_template=self.__errorlog_template,
+        error_log_name = self.__errorlog_template.instantiate(
+            module=self.module, record=input_record
         )
+        if not error_log_name:
+            return None
 
-        if not update or not error_log_name:
+        if not update:
             return error_log_name
 
         if not update_error_log_and_qc_metadata(
@@ -398,12 +395,67 @@ class CSVTransformVisitor(CSVVisitor):
 
         return error_log_name
 
+    def __get_downstream_file(
+        self, input_record: Dict[str, Any]
+    ) -> Optional[FileEntry]:
+        assert self.__project
+        visit_file_id = input_record.get("file_id")
+        if not visit_file_id:
+            return None
+
+        visit_file = self.__project.proxy.get_file(visit_file_id)
+        if not visit_file:
+            return None
+        return visit_file
+
+    def __copy_metadata(
+        self,
+        input_record: Dict[str, Any],
+        error_log_name: str,
+        downstream_gears: List[str],
+        info: FileQCModel,
+    ) -> Optional[QCStatus]:
+        visit_file = self.__get_downstream_file(input_record)
+        if not visit_file:
+            log.error(
+                "Missing file for existing visit, failed to update error log - %s",
+                error_log_name,
+            )
+            return None
+
+        if not visit_file.info_exists:
+            log.error("No QC metadata for existing visit")
+            return None
+
+        visit_file = visit_file.reload()
+        try:
+            visit_info = FileQCModel.model_validate(visit_file.info)
+        except ValidationError as error:
+            log.error("Unexpected QC metadata: %s", error)
+            return None
+
+        for ds_gear in downstream_gears:
+            ds_gear_metadata = visit_info.get(ds_gear)
+            if not ds_gear_metadata:
+                log.warning(
+                    "QC metadata not found for gear %s in the "
+                    "existing duplicate visit file %s",
+                    ds_gear,
+                    visit_file.name,
+                )
+                continue
+
+            log.info("copying metadata for gear %s", ds_gear)
+            info.set(gear_name=ds_gear, gear_model=ds_gear_metadata)
+
+        return "PASS"
+
     def __copy_downstream_gears_metadata(
         self,
         *,
         input_record: Dict[str, Any],
         downstream_gears: Optional[List[str]] = None,
-        gear_state: str = "PASS",
+        gear_state: QCStatus = "PASS",
     ) -> bool:
         """Copy any downstream gears metadata from visit file to error log
         file.
@@ -417,35 +469,37 @@ class CSVTransformVisitor(CSVVisitor):
             bool: True if copying metadata successful
         """
 
-        if not self.__project or not self.module:
-            log.warning(
-                "Parent project or module not specified to update visit error log"
-            )
+        if self.__project is None:
+            log.warning("Parent project not specified to update visit error log")
+            return False
+        if not self.module:
+            log.warning("Module not specified to update visit error log")
             return False
 
-        error_log_name = get_error_log_name(
-            module=self.module,
-            input_data=input_record,
-            errorlog_template=self.__errorlog_template,
+        error_log_name = self.__errorlog_template.instantiate(
+            module=self.module, record=input_record
         )
-
         if not error_log_name:
             return False
 
         error_log_file = self.__project.get_file(error_log_name)
         if not error_log_file:
             log.error(
-                "Failed to retrieve visit error log file %s from project",
+                "Failed to retrieve visit error log file %s/%s",
+                self.__project.label,
                 error_log_name,
             )
             return False
 
         error_log_file = error_log_file.reload()
-        info = (
-            error_log_file.info
-            if (error_log_file.info and "qc" in error_log_file.info)
-            else {"qc": {}}
-        )
+        info = FileQCModel(qc={})
+        if error_log_file.info and "qc" in error_log_file.info:
+            try:
+                info = FileQCModel.model_validate(error_log_file.info)
+            except ValidationError as error:
+                log.error(
+                    f"error reading {self.__project}/{error_log_name} metadata: {error}"
+                )
 
         # TODO: decide whether we need to show this warning, commenting out for now
         # self.__error_writer.write(
@@ -455,58 +509,29 @@ class CSVTransformVisitor(CSVVisitor):
         #                  error_type='warning'))
 
         if downstream_gears:
-            visit_file = None
-            visit_file_id = input_record.get("file_id")
-            if visit_file_id:
-                visit_file = self.__project.proxy.get_file(visit_file_id)
-            else:
-                log.error(
-                    "Missing file id for existing visit, "
-                    "failed to update error log - %s",
-                    error_log_name,
-                )
+            status = self.__copy_metadata(
+                input_record=input_record,
+                error_log_name=error_log_name,
+                downstream_gears=downstream_gears,
+                info=info,
+            )
+            if status:
+                gear_state = status
 
-            if visit_file and visit_file.info_exists:
-                visit_file = visit_file.reload()
-
-                for ds_gear in downstream_gears:
-                    ds_gear_metadata = visit_file.info.get("qc", {}).get(ds_gear, {})
-                    if not ds_gear_metadata:
-                        log.warning(
-                            "QC metadata not found for gear %s in the "
-                            "existing duplicate visit file %s",
-                            ds_gear,
-                            visit_file.name,
-                        )
-                        continue
-
-                    info["qc"][ds_gear] = ds_gear_metadata
-            else:
-                gear_state = "FAIL"
-                self.__error_writer.write(
-                    system_error(
-                        message=(
-                            "Failed to load QC metadata from existing "
-                            "duplicate visit file"
-                        ),
-                        error_type="warning",
-                    )
-                )
         else:
             log.warning(
                 "No downstream gears defined for current gear %s", self.__gear_name
             )
 
         # add current gear
-        info["qc"][self.__gear_name] = {
-            "validation": {
-                "state": gear_state.upper(),
-                "data": self.__error_writer.errors(),
-            }
-        }
+        info.set_errors(
+            gear_name=self.__gear_name,
+            status=gear_state.upper(),  # type: ignore
+            errors=self.__error_writer.errors(),
+        )
 
         try:
-            error_log_file.update_info(info)
+            error_log_file.update_info(info.model_dump(by_alias=True))
         except ApiException as error:
             log.error(error)
             return False
@@ -599,7 +624,7 @@ def run(
         input_file: the input file
         id_column: the subject identifier (usually NACCID)
         module: the module label
-        destination: Flyhweel project container
+        destination: Flywheel project container
         transformer_factory: the factory for column transformers
         preprocessor: class to run pre-processing checks
         module_configs: form ingest configs for the module